Test with embedded postgresql (#1037)
Updated some tests that depend on local PostgreSQL installation to use the postgresql-embedded (https://github.com/yandex-qatools/postgresql-embedded).

Notes:

- postgresql-embedded requires Java 8.
- These test cases are not executed in the test phase by default. To run them, override 'maven.surefire.excludeGroups' property (e.g. `./mvnw clean verify -Dmaven.surefire.excludeGroups=`).
- Travis CI runs these tests.
