--- src/test/java/org/apache/ibatis/submitted/keycolumn/MapperConfig.xml
@@ -1,41 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<!--
-
-       Copyright 2009-2016 the original author or authors.
-
-       Licensed under the Apache License, Version 2.0 (the "License");
-       you may not use this file except in compliance with the License.
-       You may obtain a copy of the License at
-
-          http://www.apache.org/licenses/LICENSE-2.0
-
-       Unless required by applicable law or agreed to in writing, software
-       distributed under the License is distributed on an "AS IS" BASIS,
-       WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-       See the License for the specific language governing permissions and
-       limitations under the License.
-
--->
-<!DOCTYPE configuration
-    PUBLIC "-//ibatis.apache.org//DTD Config 3.0//EN"
-    "http://ibatis.apache.org/dtd/ibatis-3-config.dtd">
-<configuration>
-
-  <environments default="development">
-    <environment id="development">
-      <transactionManager type="JDBC">
-        <property name="" value=""/>
-      </transactionManager>
-      <dataSource type="UNPOOLED">
-        <property name="driver" value="org.postgresql.Driver"/>
-        <property name="url" value="jdbc:postgresql:mbtest"/>
-        <property name="username" value="postgres"/>
-        <property name="password" value="root"/>
-      </dataSource>
-    </environment>
-  </environments>
-
-  <mappers>
-    <mapper resource="org/apache/ibatis/submitted/keycolumn/InsertMapper.xml"/>
-  </mappers>
-</configuration>
