--- src/test/java/org/apache/ibatis/submitted/keycolumn/setupdb.txt
@@ -1,44 +1,0 @@
-====
-       Copyright 2009-2016 the original author or authors.
-
-       Licensed under the Apache License, Version 2.0 (the "License");
-       you may not use this file except in compliance with the License.
-       You may obtain a copy of the License at
-
-          http://www.apache.org/licenses/LICENSE-2.0
-
-       Unless required by applicable law or agreed to in writing, software
-       distributed under the License is distributed on an "AS IS" BASIS,
-       WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-       See the License for the specific language governing permissions and
-       limitations under the License.
-====
-
-The tests in this package are dependent on a local install
-of PostgreSQL.  This file contains instructions for setting up
-PostgreSQL so that the tests will run successfully.  To run the tests,
-follow the database setup instructions in this file, then remove the
-@Ignore annotation on the test class and run the tests with Maven
-or a local JUnit test runner.
-
-DO NOT commit the test class without the @Ignore annotation or it will
-break the MyBatis build on machines that don't have PostgreSQL setup.
-
-1. Download and install PostgreSQL.  Set the password for the
-   "postgres" userid to "root".  If you choose a different password,
-   you will need to alter the MapperConfig.xml file accordingly.
-2. Create a database called "mbtest"
-3. Create a schema in "mbtest" database called "mbtest"
-4. Create the following table:
-
-   CREATE TABLE mbtest.test_identity
-   (
-     first_name character varying(30),
-     last_name character varying(30),
-     name_id serial NOT NULL,
-     CONSTRAINT test_identity_pkey PRIMARY KEY (name_id)
-   )
-   WITH (
-     OIDS=FALSE
-   );
-   ALTER TABLE mbtest.test_identity OWNER TO postgres;
