--- src/test/java/org/apache/ibatis/submitted/multiple_resultsets/mybatis-config.xml
@@ -1,47 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<!--
-
-       Copyright 2009-2016 the original author or authors.
-
-       Licensed under the Apache License, Version 2.0 (the "License");
-       you may not use this file except in compliance with the License.
-       You may obtain a copy of the License at
-
-          http://www.apache.org/licenses/LICENSE-2.0
-
-       Unless required by applicable law or agreed to in writing, software
-       distributed under the License is distributed on an "AS IS" BASIS,
-       WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-       See the License for the specific language governing permissions and
-       limitations under the License.
-
--->
-<!DOCTYPE configuration
-    PUBLIC "-//mybatis.org//DTD Config 3.0//EN"
-    "http://mybatis.org/dtd/mybatis-3-config.dtd">
-
-<configuration>
-
-	<settings>
-		<setting name="mapUnderscoreToCamelCase" value="true" />
-	</settings>
-
-	<environments default="development">
-		<environment id="development">
-			<transactionManager type="JDBC">
-				<property name="" value="" />
-			</transactionManager>
-			<dataSource type="UNPOOLED">
-		        <property name="driver" value="org.postgresql.Driver"/>
-		        <property name="url" value="jdbc:postgresql:mbtest"/>
-		        <property name="username" value="postgres"/>
-		        <property name="password" value="root"/>
-			</dataSource>
-		</environment>
-	</environments>
-
-	<mappers>
-		<mapper resource="org/apache/ibatis/submitted/multiple_resultsets/Mapper.xml" />
-	</mappers>
-
-</configuration>
