+++ src/test/java/org/apache/ibatis/submitted/usesjava8/keycolumn/InsertMapper.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8" ?>
 <!--
 
-       Copyright 2009-2016 the original author or authors.
+       Copyright 2009-2017 the original author or authors.
 
        Licensed under the Apache License, Version 2.0 (the "License");
        you may not use this file except in compliance with the License.
@@ -19,8 +19,8 @@
 <!DOCTYPE mapper
     PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
     "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="org.apache.ibatis.submitted.keycolumn.InsertMapper" >
-  <insert id="insertNameMapped" parameterType="org.apache.ibatis.submitted.keycolumn.Name" useGeneratedKeys="true" keyProperty="id" keyColumn="name_id">
+<mapper namespace="org.apache.ibatis.submitted.usesjava8.keycolumn.InsertMapper" >
+  <insert id="insertNameMapped" parameterType="org.apache.ibatis.submitted.usesjava8.keycolumn.Name" useGeneratedKeys="true" keyProperty="id" keyColumn="name_id">
     insert into mbtest.test_identity
       (first_name, last_name)
     values(#{firstName,jdbcType=VARCHAR}, #{lastName,jdbcType=VARCHAR})
