+++ src/test/java/org/apache/ibatis/submitted/usesjava8/keycolumn/InsertTest.java
@@ -1,5 +1,5 @@
 /**
- *    Copyright 2009-2015 the original author or authors.
+ *    Copyright 2009-2017 the original author or authors.
  *
  *    Licensed under the Apache License, Version 2.0 (the "License");
  *    you may not use this file except in compliance with the License.
@@ -13,42 +13,76 @@
  *    See the License for the specific language governing permissions and
  *    limitations under the License.
  */
-package org.apache.ibatis.submitted.keycolumn;
+package org.apache.ibatis.submitted.usesjava8.keycolumn;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
 
 import java.io.Reader;
+import java.nio.file.Paths;
+import java.sql.Connection;
+import java.util.Collections;
 import java.util.List;
 
+import org.apache.ibatis.datasource.unpooled.UnpooledDataSource;
 import org.apache.ibatis.executor.BatchResult;
 import org.apache.ibatis.io.Resources;
+import org.apache.ibatis.jdbc.ScriptRunner;
+import org.apache.ibatis.mapping.Environment;
+import org.apache.ibatis.session.Configuration;
 import org.apache.ibatis.session.ExecutorType;
 import org.apache.ibatis.session.SqlSession;
 import org.apache.ibatis.session.SqlSessionFactory;
 import org.apache.ibatis.session.SqlSessionFactoryBuilder;
-import org.junit.Ignore;
+import org.apache.ibatis.test.EmbeddedPostgresqlTests;
+import org.apache.ibatis.transaction.jdbc.JdbcTransactionFactory;
+import org.junit.AfterClass;
+import org.junit.BeforeClass;
 import org.junit.Test;
+import org.junit.experimental.categories.Category;
 
-/*
- * This class contains tests for issue #84 - where the Jdbc3KeyGenerator
- * needs to know the key column name because it is not the first column
- * in the resultset returned from getGeneratedKeys().
- *
- * If PostgreSQL is setup as described in setupdb.txt, then remove
- * the @Ignore annotation to enable the tests.
+import ru.yandex.qatools.embed.postgresql.EmbeddedPostgres;
+import ru.yandex.qatools.embed.postgresql.util.SocketUtil;
 
+/**
  * @author Jeff Butler
  */
-@Ignore("See setupdb.txt for instructions on how to run the tests in this class")
+@Category(EmbeddedPostgresqlTests.class)
 public class InsertTest {
+
+  private static final EmbeddedPostgres postgres = new EmbeddedPostgres();
+
+  private static SqlSessionFactory sqlSessionFactory;
+
+  @BeforeClass
+  public static void setUp() throws Exception {
+    // Launch PostgreSQL server. Download / unarchive if necessary.
+    String url = postgres.start(EmbeddedPostgres.cachedRuntimeConfig(Paths.get("target/postgres")), "localhost", SocketUtil.findFreePort(), "keycolumn", "postgres", "root", Collections.emptyList());
+
+    Configuration configuration = new Configuration();
+    Environment environment = new Environment("development", new JdbcTransactionFactory(), new UnpooledDataSource(
+        "org.postgresql.Driver", url, null));
+    configuration.setEnvironment(environment);
+    configuration.addMapper(InsertMapper.class);
+    sqlSessionFactory = new SqlSessionFactoryBuilder().build(configuration);
+
+    try (SqlSession session = sqlSessionFactory.openSession();
+        Connection conn = session.getConnection();
+        Reader reader = Resources.getResourceAsReader("org/apache/ibatis/submitted/usesjava8/keycolumn/CreateDB.sql")) {
+      ScriptRunner runner = new ScriptRunner(conn);
+      runner.setLogWriter(null);
+      runner.runScript(reader);
+    }
+  }
+
+  @AfterClass
+  public static void tearDown() {
+    postgres.stop();
+  }
 
   @Test
   public void testInsertAnnotated() throws Exception {
-    Reader reader = Resources.getResourceAsReader("org/apache/ibatis/submitted/keycolumn/MapperConfig.xml");
-    SqlSessionFactory sqlSessionFactory = new SqlSessionFactoryBuilder().build(reader);
-    SqlSession sqlSession = sqlSessionFactory.openSession();
-    try {
+    try (SqlSession sqlSession = sqlSessionFactory.openSession()) {
       InsertMapper mapper = sqlSession.getMapper(InsertMapper.class);
       Name name = new Name();
       name.setFirstName("Fred");
@@ -58,17 +92,12 @@ public class InsertTest {
 
       assertNotNull(name.getId());
       assertEquals(1, rows);
-    } finally {
-      sqlSession.close();
     }
   }
 
   @Test
   public void testInsertMapped() throws Exception {
-    Reader reader = Resources.getResourceAsReader("org/apache/ibatis/submitted/keycolumn/MapperConfig.xml");
-    SqlSessionFactory sqlSessionFactory = new SqlSessionFactoryBuilder().build(reader);
-    SqlSession sqlSession = sqlSessionFactory.openSession();
-    try {
+    try (SqlSession sqlSession = sqlSessionFactory.openSession()) {
       InsertMapper mapper = sqlSession.getMapper(InsertMapper.class);
       Name name = new Name();
       name.setFirstName("Fred");
@@ -78,18 +107,12 @@ public class InsertTest {
 
       assertNotNull(name.getId());
       assertEquals(1, rows);
-    } finally {
-      sqlSession.close();
     }
   }
 
-  @Ignore // Not supported yet in PostgreSQL
   @Test
   public void testInsertMappedBatch() throws Exception {
-    Reader reader = Resources.getResourceAsReader("org/apache/ibatis/submitted/keycolumn/MapperConfig.xml");
-    SqlSessionFactory sqlSessionFactory = new SqlSessionFactoryBuilder().build(reader);
-    SqlSession sqlSession = sqlSessionFactory.openSession(ExecutorType.BATCH);
-    try {
+    try (SqlSession sqlSession = sqlSessionFactory.openSession(ExecutorType.BATCH)) {
       InsertMapper mapper = sqlSession.getMapper(InsertMapper.class);
       Name name = new Name();
       name.setFirstName("Fred");
@@ -103,8 +126,6 @@ public class InsertTest {
       assertNotNull(name.getId());
       assertNotNull(name2.getId());
       assertEquals(1, batchResults.size());
-    } finally {
-      sqlSession.close();
     }
   }
