+++ src/test/java/org/apache/ibatis/submitted/usesjava8/multiple_resultsets/Mapper.java
@@ -1,5 +1,5 @@
 /**
- *    Copyright 2009-2015 the original author or authors.
+ *    Copyright 2009-2017 the original author or authors.
  *
  *    Licensed under the Apache License, Version 2.0 (the "License");
  *    you may not use this file except in compliance with the License.
@@ -13,27 +13,12 @@
  *    See the License for the specific language governing permissions and
  *    limitations under the License.
  */
-package org.apache.ibatis.submitted.multiple_resultsets;
+package org.apache.ibatis.submitted.usesjava8.multiple_resultsets;
 
-public class OrderHeader {
+import java.util.List;
 
-  private int orderId;
-  private String custName;
+public interface Mapper {
 
-  public int getOrderId() {
-    return orderId;
-  }
-
-  public void setOrderId(int orderId) {
-    this.orderId = orderId;
-  }
-
-  public String getCustName() {
-    return custName;
-  }
-
-  public void setCustName(String custName) {
-    this.custName = custName;
-  }
+  List<?> getUsersAndGroups();
 
 }
