+++ src/test/java/org/apache/ibatis/submitted/usesjava8/multiple_resultsets/OrderHeader.java
@@ -1,0 +1,39 @@
+/**
+ *    Copyright 2009-2017 the original author or authors.
+ *
+ *    Licensed under the Apache License, Version 2.0 (the "License");
+ *    you may not use this file except in compliance with the License.
+ *    You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *    Unless required by applicable law or agreed to in writing, software
+ *    distributed under the License is distributed on an "AS IS" BASIS,
+ *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *    See the License for the specific language governing permissions and
+ *    limitations under the License.
+ */
+package org.apache.ibatis.submitted.usesjava8.multiple_resultsets;
+
+public class OrderHeader {
+
+  private int orderId;
+  private String custName;
+
+  public int getOrderId() {
+    return orderId;
+  }
+
+  public void setOrderId(int orderId) {
+    this.orderId = orderId;
+  }
+
+  public String getCustName() {
+    return custName;
+  }
+
+  public void setCustName(String custName) {
+    this.custName = custName;
+  }
+
+}
