+++ src/test/java/org/apache/ibatis/submitted/usesjava8/refcursor/Order.java
@@ -1,5 +1,5 @@
 /**
- *    Copyright 2009-2015 the original author or authors.
+ *    Copyright 2009-2017 the original author or authors.
  *
  *    Licensed under the Apache License, Version 2.0 (the "License");
  *    you may not use this file except in compliance with the License.
@@ -13,7 +13,7 @@
  *    See the License for the specific language governing permissions and
  *    limitations under the License.
  */
-package org.apache.ibatis.submitted.refcursor;
+package org.apache.ibatis.submitted.usesjava8.refcursor;
 
 import java.util.List;
