+++ src/test/java/org/apache/ibatis/submitted/usesjava8/refcursor/OrdersMapper.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!--
 
-       Copyright 2009-2016 the original author or authors.
+       Copyright 2009-2017 the original author or authors.
 
        Licensed under the Apache License, Version 2.0 (the "License");
        you may not use this file except in compliance with the License.
@@ -16,12 +16,15 @@
        limitations under the License.
 
 -->
-<!DOCTYPE mapper PUBLIC "-//ibatis.apache.org//DTD Mapper 3.0//EN" "http://ibatis.apache.org/dtd/ibatis-3-mapper.dtd">
-<mapper namespace="org.apache.ibatis.submitted.refcursor.OrdersMapper">
-  <resultMap type="org.apache.ibatis.submitted.refcursor.Order" id="OrderResult">
+<!DOCTYPE mapper
+    PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
+    "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
+
+<mapper namespace="org.apache.ibatis.submitted.usesjava8.refcursor.OrdersMapper">
+  <resultMap type="org.apache.ibatis.submitted.usesjava8.refcursor.Order" id="OrderResult">
     <id property="orderId" column="ORDER_ID"/>
     <result property="customerName" column="CUST_NAME"/>
-    <collection property="detailLines" ofType="org.apache.ibatis.submitted.refcursor.OrderDetail">
+    <collection property="detailLines" ofType="org.apache.ibatis.submitted.usesjava8.refcursor.OrderDetail">
       <id property="lineNumber" column="LINE_NUMBER"/>
       <result property="quantity" column="QUANTITY"/>
       <result property="description" column="ITEM_DESCRIPTION"/>
@@ -34,7 +37,7 @@
   </update>
 
   <update id="getOrder2" parameterType="map" statementType="CALLABLE">
-    { #{order,jdbcType=OTHER,mode=OUT,resultMap=org.apache.ibatis.submitted.refcursor.OrdersMapper.OrderResult,javaType=java.sql.ResultSet} = 
+    { #{order,jdbcType=OTHER,mode=OUT,resultMap=org.apache.ibatis.submitted.usesjava8.refcursor.OrdersMapper.OrderResult,javaType=java.sql.ResultSet} = 
         call mbtest.get_order(#{orderId,jdbcType=INTEGER,mode=IN}) }
   </update>
 </mapper>
