+++ src/test/java/org/apache/ibatis/test/SlowTests.java
@@ -1,5 +1,5 @@
 /**
- *    Copyright 2009-2015 the original author or authors.
+ *    Copyright 2009-2017 the original author or authors.
  *
  *    Licensed under the Apache License, Version 2.0 (the "License");
  *    you may not use this file except in compliance with the License.
@@ -13,12 +13,8 @@
  *    See the License for the specific language governing permissions and
  *    limitations under the License.
  */
-package org.apache.ibatis.submitted.multiple_resultsets;
 
-import java.util.List;
+package org.apache.ibatis.test;
 
-public interface Mapper {
-
-  List<?> getUsersAndGroups();
-
+public interface SlowTests {
 }
