+++ travis/install.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 #
-#    Copyright 2009-2016 the original author or authors.
+#    Copyright 2009-2017 the original author or authors.
 #
 #    Licensed under the Apache License, Version 2.0 (the "License");
 #    you may not use this file except in compliance with the License.
